backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymupdf
python-dateutil
